+++ recipes/VideoBots.py
@@ -5,8 +5,6 @@ from itertools import zip_longest
 
 import gooey_gui as gui
 import typing_extensions
-from django.db.models import Q, QuerySet
-from furl import furl
 from pydantic import BaseModel, Field
 
 from ai_models.llm_openapi import LLMMarker
@@ -17,11 +15,9 @@ from bots.models import (
     PublishedRun,
     SavedRun,
     Workflow,
-    WorkflowAccessLevel,
 )
-from celeryapp.tasks import send_integration_attempt_email
 from daras_ai.image_input import truncate_text_words
-from daras_ai_v2 import icons, settings
+from daras_ai_v2 import settings
 from daras_ai_v2.asr import (
     AsrModels,
     TranslationModels,
@@ -38,14 +34,9 @@ from daras_ai_v2.azure_doc_extract import (
     azure_form_recognizer,
 )
 from daras_ai_v2.base import BasePage, RecipeRunState, RecipeTabs, StateKeys
-from daras_ai_v2.bot_integration_connect import connect_bot_to_published_run
-from daras_ai_v2.bot_integration_widgets import (
-    broadcast_input,
-    general_integration_settings,
-    integrations_welcome_screen,
-    web_widget_config,
-)
+from daras_ai_v2.bot_integration_widgets import integrations_welcome_screen
 from daras_ai_v2.csv_lines import csv_decode_row
+from daras_ai_v2.integrations_tab import render_integrations_tab
 from daras_ai_v2.doc_search_settings_widgets import (
     SUPPORTED_SPREADSHEET_TYPES,
     bulk_documents_uploader,
@@ -111,7 +102,6 @@ from functions.base_llm_tool import (
     get_tool_from_call,
 )
 from functions.workflow_tools import DynamicLLMToolLoader
-from payments.plans import PricingPlan
 from recipes.DocExtract import document_intelligence_settings
 from recipes.DocSearch import get_top_k_references, references_as_prompt
 from recipes.GoogleGPT import SearchReference
@@ -1661,600 +1651,24 @@ if (typeof GooeyEmbed !== "undefined" && GooeyEmbed.copilotPreviewControl) {
         return tabs
 
     def render_selected_tab(self):
-        super().render_selected_tab()
-
         if self.tab == RecipeTabs.integrations:
-            self.render_integrations_tab()
-
-    def render_integrations_tab(self):
-        from daras_ai_v2.breadcrumbs import get_title_breadcrumbs
-
-        gui.newline()
-
-        # not signed in case
-        if not self.request.user or self.request.user.is_anonymous:
-            integrations_welcome_screen(title="Connect your Copilot")
-            gui.newline()
-            with gui.center():
-                gui.anchor("Get Started", href=self.get_auth_url(), type="primary")
-            return
-
-        sr, pr = self.current_sr_pr
-
-        # make user the user knows that they are on a saved run not the published run
-        if pr and pr.saved_run_id != sr.id:
-            last_saved_url = self.app_url(
-                tab=RecipeTabs.integrations, example_id=pr.published_run_id
-            )
-            gui.caption(
-                f"Note: You seem to have unpublished changes. Deployments use the [last saved version]({last_saved_url}), not the currently visible edits.",
-                className="text-center text-muted",
-            )
-
-        # see which integrations are available to the user for the published run
-        integrations_q = Q(published_run=pr) | Q(
-            saved_run__example_id=pr.published_run_id
-        )
-        if not self.is_current_user_admin():
-            integrations_q &= Q(workspace=self.current_workspace)
-
-        integrations_qs: QuerySet[BotIntegration] = BotIntegration.objects.filter(
-            integrations_q
-        ).order_by("platform", "-created_at")
-
-        run_title = get_title_breadcrumbs(VideoBotsPage, sr, pr).title_with_prefix()
-
-        # no connected integrations on this run
-        if not (integrations_qs and integrations_qs.exists()):
-            self.render_integrations_add(
-                label="#### Connect your Copilot",
-                run_title=run_title,
-                pr=pr,
-            )
-            return
-
-        # this gets triggered on the /add route
-        if gui.session_state.pop("--add-integration", None):
-            self.render_integrations_add(
-                label="#### Deploy to a New Channel",
-                run_title=run_title,
-                pr=pr,
-            )
-            with gui.center():
-                if gui.button("Return to Configure"):
-                    cancel_url = self.current_app_url(RecipeTabs.integrations)
-                    raise gui.RedirectException(cancel_url)
-            return
-
-        with gui.center():
-            # signed in, can edit, and has connected botintegrations on this run
-            self.render_integrations_settings(
-                integrations=list(integrations_qs), run_title=run_title
-            )
-
-    def render_integrations_add(self, label: str, run_title: str, pr: PublishedRun):
-        from routers.facebook_api import fb_connect_url
-        from routers.slack_api import slack_connect_url
-        from routers.facebook_api import wa_connect_url
-        from number_cycling.utils import create_bot_integration_with_extension
-        from number_cycling.models import SharedPhoneNumber
-        from daras_ai_v2.bot_integration_widgets import (
-            render_telegram_connect_dialog,
-        )
-
-        gui.write(label, unsafe_allow_html=True, className="text-center")
-
-        can_edit = (
-            self.request.user
-            and WorkflowAccessLevel.can_user_edit_published_run(
-                workspace=self.current_workspace,
-                user=self.request.user,
-                pr=pr,
-            )
-        )
-
-        telegram_dialog = gui.use_confirm_dialog(
-            key="telegram-connect", close_on_confirm=False
-        )
-
-        gui.newline()
-
-        pressed_platform = None
-        country_code = ""
-        with (
-            gui.tag("table", className="d-flex justify-content-center"),
-            gui.tag("tbody"),
-        ):
-            for choice in connect_choices:
-                with gui.tag("tr"):
-                    with gui.tag("td"):
-                        if gui.button(
-                            f'<img src="{choice.img}" alt="{choice.platform.name}" style="max-width: 80%; max-height: 90%" draggable="false">',
-                            className="p-0 border border-1 border-secondary rounded",
-                            style=dict(width="160px", height="60px"),
-                        ):
-                            pressed_platform = choice.platform
-                    with gui.tag("td", className="ps-3"):
-                        if choice.platform == Platform.TWILIO:
-                            country_code = render_twilio_label(choice.label)
-                        else:
-                            gui.caption(choice.label)
-
-        if not can_edit:
-            gui.caption(
-                "P.S. You're not an owner of this saved workflow, so we'll create a copy of it in your Saved Runs.",
-                className="text-center text-muted",
-            )
-
-        if pressed_platform:
-            if not can_edit and pressed_platform != Platform.TELEGRAM:
-                run_title = f"{self.request.user and self.request.user.first_name_possesive()} {run_title}"
-                pr = pr.duplicate(
-                    user=self.request.user,
-                    workspace=self.current_workspace,
-                    title=run_title,
-                    notes=pr.notes,
-                )
-
-            match pressed_platform:
-                case Platform.TELEGRAM:
-                    telegram_dialog.set_open(True)
-                case Platform.WEB:
-                    bi = BotIntegration.objects.create(
-                        name=run_title,
-                        created_by=self.request.user,
-                        workspace=self.current_workspace,
-                        platform=Platform.WEB,
-                    )
-                    redirect_url = connect_bot_to_published_run(bi, pr, overwrite=True)
-                case Platform.WHATSAPP:
-                    try:
-                        bi = create_bot_integration_with_extension(
-                            name=run_title,
-                            created_by=self.request.user,
-                            workspace=self.current_workspace,
-                            platform=Platform.WHATSAPP,
-                        )
-                        redirect_url = connect_bot_to_published_run(
-                            bi, pr, overwrite=True
-                        )
-                    except SharedPhoneNumber.DoesNotExist:
-                        redirect_url = wa_connect_url(pr.id)
-                case Platform.SLACK:
-                    redirect_url = slack_connect_url(pr.id)
-                case Platform.FACEBOOK:
-                    redirect_url = fb_connect_url(pr.id)
-                case Platform.TWILIO:
-                    try:
-                        bi = create_bot_integration_with_extension(
-                            name=run_title,
-                            created_by=self.request.user,
-                            workspace=self.current_workspace,
-                            platform=Platform.TWILIO,
-                            country_code=country_code,
-                        )
-                    except SharedPhoneNumber.DoesNotExist as e:
-                        gui.caption(f"{e}", className="text-center text-danger")
-                        return
-
-                    redirect_url = connect_bot_to_published_run(bi, pr, overwrite=True)
-                case _:
-                    raise ValueError(f"Unsupported platform: {pressed_platform}")
-
-            if pressed_platform != Platform.TELEGRAM:
-                if not self.is_current_user_admin():
-                    send_integration_attempt_email.delay(
-                        user_id=self.request.user.id,
-                        platform=pressed_platform,
-                        run_url=self.current_app_url() or "",
-                    )
-                raise gui.RedirectException(redirect_url)
-
-        if telegram_dialog.is_open:
-            render_telegram_connect_dialog(
-                telegram_dialog,
-                pr=pr,
-                run_title=run_title,
-                can_edit=can_edit,
-                user=self.request.user,
-                workspace=self.current_workspace,
-                is_admin=self.is_current_user_admin(),
-                current_app_url=self.current_app_url() or "",
-            )
-
-        gui.newline()
-        api_tab_url = self.current_app_url(RecipeTabs.run_as_api)
-        gui.write(
-            f"Or use [our API]({api_tab_url}) to build custom integrations with your server.",
-            className="text-center",
-        )
-
-    def render_integrations_settings(
-        self, integrations: list[BotIntegration], run_title: str
-    ):
-        from daras_ai_v2.copy_to_clipboard_button_widget import copy_to_clipboard_button
-        from routers.facebook_api import wa_connect_url
-        from recipes.QRCodeGenerator import generate_qr_code
-
-        with gui.div(className="w-100 text-start"):
-            col1, col2 = gui.columns(2)
-
-        with col1:
-            gui.markdown("#### Configure your Copilot")
-            gui.newline()
-
-            if len(integrations) > 1:
-                with gui.div(
-                    style={"width": "100%", "maxWidth": "500px", "textAlign": "left"}
-                ):
-                    integrations_map = {i.id: i for i in integrations}
-                    bi_id = gui.selectbox(
-                        label="",
-                        options=integrations_map.keys(),
-                        format_func=lambda bi_id: f"{Platform(integrations_map[bi_id].platform).get_icon()} &nbsp; {integrations_map[bi_id].name}",
-                        key="bi_id",
-                    )
-                    bi = integrations_map[bi_id]
-                    old_bi_id = gui.session_state.get("old_bi_id", bi_id)
-                    if bi_id != old_bi_id:
-                        raise gui.RedirectException(
-                            self.current_app_url(
-                                RecipeTabs.integrations,
-                                path_params=dict(
-                                    integration_id=bi.api_integration_id()
-                                ),
-                            )
-                        )
-                    gui.session_state["old_bi_id"] = bi_id
-            else:
-                bi = integrations[0]
-
-        with col2:
-            test_link = bi.get_bot_test_link()
-            if bi.demo_qr_code_image:
-                img_src = bi.demo_qr_code_image
-            elif test_link:
-                img_src = generate_qr_code(test_link)
-            else:
-                img_src = None
-
-            if img_src is not None:
-                gui.image(img_src, style={"maxWidth": "125px", "maxHeight": "125px"})
-
-        if bi.platform == Platform.WEB:
-            web_widget_config(
-                bi=bi,
-                user=self.request.user,
-                hostname=self.request.url and self.request.url.hostname,
-            )
-            with gui.div(className="w-100"):
-                gui.write("---")
-
-        icon = Platform(bi.platform).get_icon()
-        with gui.div(className="w-100 text-start"):
-            col1, col2 = gui.columns(2, style={"alignItems": "center"})
-            with col1:
-                if bi.extension_number:
-                    gui.write("###### Connected to Extension")
-                else:
-                    gui.write("###### Connected to")
-
-                gui.write(f"{icon} {bi}", unsafe_allow_html=True)
-            with col2:
-                if not test_link:
-                    gui.write("Message quicklink not available.")
-                elif bi.platform == Platform.TWILIO:
-                    copy_to_clipboard_button(
-                        f"{icons.copy_solid} Copy Phone Number",
-                        value=test_link.lstrip("tel:"),
-                        type="secondary",
-                    )
-                else:
-                    copy_to_clipboard_button(
-                        f"{icons.copy_solid} Copy {Platform(bi.platform).label} Link",
-                        value=test_link,
-                        type="secondary",
-                    )
-
-                if bi.platform == Platform.FACEBOOK:
-                    gui.anchor(
-                        '<i class="fa-regular fa-inbox"></i> Open Inbox',
-                        "https://www.facebook.com/latest/inbox",
-                        unsafe_allow_html=True,
-                        new_tab=True,
-                    )
-
-            col1, col2 = gui.columns(2, style={"alignItems": "center"})
-            with col1:
-                gui.write("###### Test")
-                if bi.platform == Platform.TWILIO:
-                    test_caption = f"Call or send a text message via {Platform(bi.platform).label}."
-                else:
-                    test_caption = f"Send a message via {Platform(bi.platform).label}."
-                if bi.extension_number:
-                    test_caption += f" (with extension {bi.extension_number})."
-                help_text = None
-                if bi.platform == Platform.TWILIO:
-                    help_text = (
-                        "**SMS:** Send `/extension <extension number>` to connect to the agent. `/disconnect` to start fresh.\n\n"
-                        "**Voice Call:** ` * <extension number>` to change extension, `*#` to disconnect."
-                    )
-                gui.caption(test_caption, help=help_text)
-            with col2:
-                if not test_link:
-                    gui.write("Message quicklink not available.")
-                elif bi.platform == Platform.FACEBOOK:
-                    gui.anchor(
-                        f"{icon} Open Profile",
-                        test_link,
-                        unsafe_allow_html=True,
-                        new_tab=True,
-                    )
-                elif bi.platform == Platform.TWILIO:
-                    gui.anchor(
-                        '<i class="fa-regular fa-phone"></i> Start Voice Call',
-                        test_link,
-                        unsafe_allow_html=True,
-                        new_tab=True,
-                    )
-                    sms_url = furl("sms:") / bi.twilio_phone_number.as_e164
-                    if bi.extension_number:
-                        sms_url.args["body"] = f"{bi.extension_number}"
-                    gui.anchor(
-                        '<i class="fa-regular fa-sms"></i> Send SMS',
-                        str(sms_url),
-                        unsafe_allow_html=True,
-                        new_tab=True,
-                    )
-                else:
-                    gui.anchor(
-                        f"{icon} Message {bi.get_display_name()}",
-                        test_link,
-                        unsafe_allow_html=True,
-                        new_tab=True,
-                    )
-
-            if bi.platform == Platform.TELEGRAM:
-                col1, col2 = gui.columns(2, style={"alignItems": "center"})
-                with col1:
-                    gui.write("###### Manage Bot")
-                    gui.caption(
-                        "Open BotFather in Telegram to manage your bot's name, photo, and other settings."
-                    )
-                with col2:
-                    telegram_botfather_url = furl(
-                        "https://t.me/BotFather", query_params={"text": "/mybots"}
-                    )
-                    gui.anchor(
-                        f"{icon} Open BotFather",
-                        str(telegram_botfather_url),
-                        unsafe_allow_html=True,
-                        new_tab=True,
-                        type="secondary",
-                    )
-
-            if bi.platform == Platform.WHATSAPP and bi.extension_number:
-                is_enterprise = (
-                    self.current_workspace.subscription
-                    and PricingPlan.from_sub(self.current_workspace.subscription)
-                    == PricingPlan.ENTERPRISE
-                )
-
-                col1, col2 = gui.columns(2, style={"alignItems": "center"})
-                with col1:
-                    gui.write("###### Bring your own number")
-                    gui.write(
-                        "Connect your mobile # (that's not already on WhatsApp) with your Facebook Business Profile. [Help Guide](https://gooey.ai/docs/guides/copilot/deploy-to-whatsapp)",
-                    )
-
-                with col2:
-                    gui.anchor(
-                        "Connect number",
-                        href=wa_connect_url(self.current_pr.id),
-                        style={
-                            "backgroundColor": "#1877F2",
-                            "color": "white",
-                            "width": "100%",
-                            "maxWidth": "225px",
-                        },
-                        type="secondary",
-                    )
-
-                gui.html("""
-                        <div class="d-flex align-items-center my-2">
-                            <hr class="flex-grow-1">
-                            <span class="px-3 text-muted">or</span>
-                            <hr class="flex-grow-1">
-                        </div>
-                        """)
-
-                col1, col2 = gui.columns(2, style={"alignItems": "center"})
-                with col1:
-                    gui.write("###### Buy a dedicated number")
-                    # Check if current workspace has enterprise subscription
-                    if is_enterprise:
-                        gui.write(
-                            "As a premium customer, please contact us to setup a managed number"
-                        )
-                    else:
-                        gui.write(
-                            f"[Upgrade]({settings.PRICING_DETAILS_URL}) for a number managed by Gooey.AI"
-                        )
-                with col2:
-                    if is_enterprise:
-                        gui.anchor(
-                            "Contact",
-                            href=settings.CONTACT_URL,
-                            style={
-                                "width": "100%",
-                                "maxWidth": "225px",
-                            },
-                            type="primary",
-                        )
-                    else:
-                        gui.anchor(
-                            "Upgrade",
-                            href=settings.PRICING_DETAILS_URL,
-                            style={
-                                "width": "100%",
-                                "maxWidth": "225px",
-                            },
-                            type="primary",
-                        )
-
-                gui.write("---")
-
-            if bi.platform == Platform.TWILIO and bi.extension_number:
-                col1, col2 = gui.columns(2, style={"alignItems": "center"})
-                is_enterprise = (
-                    self.current_workspace.subscription
-                    and PricingPlan.from_sub(self.current_workspace.subscription)
-                    == PricingPlan.ENTERPRISE
-                )
-                with col1:
-                    gui.write("###### Get a Dedicated Number")
-                    if is_enterprise:
-                        gui.write(
-                            "As a premium customer, please contact us to set up a managed number"
-                        )
-                    else:
-                        gui.write(
-                            f"[Upgrade]({settings.PRICING_DETAILS_URL}) for a dedicated number managed by Gooey.AI"
-                        )
-                with col2:
-                    if is_enterprise:
-                        gui.anchor(
-                            "Contact",
-                            href=settings.CONTACT_URL,
-                            style={
-                                "width": "100%",
-                                "maxWidth": "225px",
-                            },
-                            type="primary",
-                        )
-                    else:
-                        gui.anchor(
-                            "Upgrade",
-                            href=settings.PRICING_DETAILS_URL,
-                            style={
-                                "width": "100%",
-                                "maxWidth": "225px",
-                            },
-                            type="primary",
-                        )
-            col1, col2 = gui.columns(2, style={"alignItems": "center"})
-            with col1:
-                gui.write("###### Understand your Users")
-                gui.caption("See real-time analytics.")
-            with col2:
-                gui.anchor(
-                    "📊 View Analytics",
-                    str(
-                        furl(
-                            self.current_app_url(
-                                RecipeTabs.integrations,
-                                path_params=dict(
-                                    integration_id=bi.api_integration_id()
-                                ),
-                            )
-                        )
-                        / "stats/"
-                    ),
-                    new_tab=True,
-                )
-                if (
-                    bi.platform == Platform.TWILIO
-                    and bi.twilio_phone_number_sid
-                    and not bi.extension_number
-                ):
-                    gui.anchor(
-                        f"{icon} Open Twilio Console",
-                        str(
-                            furl(
-                                "https://console.twilio.com/us1/develop/phone-numbers/manage/incoming/"
-                            )
-                            / bi.twilio_phone_number_sid
-                            / "calls"
-                        ),
-                        unsafe_allow_html=True,
-                        new_tab=True,
-                    )
-
-            if bi.platform == Platform.WHATSAPP and bi.wa_business_waba_id:
-                col1, col2 = gui.columns(2, style={"alignItems": "center"})
-                with col1:
-                    gui.write("###### WhatsApp Business Management")
-                    gui.caption(
-                        "Access your WhatsApp account on Meta to approve message templates, etc."
-                    )
-                with col2:
-                    gui.anchor(
-                        "Business Settings",
-                        str(
-                            furl(
-                                "https://business.facebook.com/settings/whatsapp-business-accounts/"
-                            )
-                            / bi.wa_business_waba_id
-                        ),
-                        new_tab=True,
-                    )
-                    gui.anchor(
-                        "WhatsApp Manager",
-                        str(
-                            furl(
-                                "https://business.facebook.com/wa/manage/home/",
-                                query_params=dict(waba_id=bi.wa_business_waba_id),
-                            )
-                        ),
-                        new_tab=True,
-                    )
-
-            col1, col2 = gui.columns(2, style={"alignItems": "center"})
-            with col1:
-                gui.write("###### Add Deployment")
-                gui.caption(f"Add another connection for {run_title}.")
-            with col2:
-                gui.anchor(
-                    f'<img align="left" width="24" height="24" src="{icons.integrations_img}"> &nbsp; Add Deployment',
-                    str(furl(self.current_app_url(RecipeTabs.integrations)) / "add/"),
-                    unsafe_allow_html=True,
-                )
-
-            gui.write("---")
-            gui.newline()
-            general_integration_settings(
-                user=self.request.user,
-                workspace=self.current_workspace,
-                bi=bi,
-                has_test_link=bool(test_link),
-            )
-            gui.write("---")
-
-            if bi.platform in [Platform.SLACK, Platform.WHATSAPP, Platform.TWILIO]:
+            user = self.request.user
+            # not signed in case
+            if not user or user.is_anonymous:
+                integrations_welcome_screen(title="Connect your Copilot")
                 gui.newline()
-                broadcast_input(bi)
-                gui.write("---")
-
-            col1, col2 = gui.columns(2, style={"alignItems": "center"})
-            with col1:
-                gui.write("###### Disconnect")
-                gui.caption(
-                    f"Disconnect {run_title} from {Platform(bi.platform).label} {bi.get_display_name()}."
-                )
-            with col2:
-                if gui.button(
-                    "💔️ Disconnect",
-                    key="btn_disconnect",
-                ):
-                    if bi.platform == Platform.TELEGRAM:
-                        bi.telegram_bot_token = ""
-                        bi.telegram_bot_id = None
-                        bi.telegram_bot_user_name = ""
-
-                    bi.saved_run = None
-                    bi.published_run = None
-                    bi.save()
-                    gui.rerun()
+                with gui.center():
+                    gui.anchor("Get Started", href=self.get_auth_url(), type="primary")
+                return
+            sr, pr = self.current_sr_pr
+            render_integrations_tab(
+                user=self.request.user,
+                workspace=self.current_workspace,
+                saved_run=sr,
+                published_run=pr,
+            )
+        else:
+            super().render_selected_tab()
 
 
 def messages_as_prompt(query_msgs: list[dict]) -> str:
@@ -2295,66 +1709,3 @@ def convo_window_clipper(
         if calc_gpt_tokens(window[i:]) > max_tokens:
             return i + step
     return 0
-
-
-class ConnectChoice(typing.NamedTuple):
-    platform: Platform
-    img: str
-    label: str
-
-
-def render_twilio_label(default_label: str = ""):
-    from number_cycling.models import SharedPhoneNumber
-    from number_cycling.utils import country_code_label
-
-    country_codes = SharedPhoneNumber.objects.available_country_codes(Platform.TWILIO)
-    if not country_codes:
-        gui.caption(default_label)
-        return ""
-
-    with gui.div(className="d-flex align-items-center flex-wrap gap-1 text-muted"):
-        gui.caption("Instantly get a")
-        selected_country_code = gui.selectbox(
-            label="",
-            options=country_codes,
-            format_func=country_code_label,
-            key="_shared_number_country",
-            label_visibility="collapsed",
-        )
-        gui.caption("number to call or text.")
-
-    return selected_country_code or ""
-
-
-connect_choices = [
-    ConnectChoice(
-        platform=Platform.WEB,
-        img="https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/29c0ffa8-29d8-11f1-ae3c-02420a00014a/image_400x400.png",
-        label="Connect to your own App or Website.",
-    ),
-    ConnectChoice(
-        platform=Platform.WHATSAPP,
-        img="https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/1e49ad50-d6c9-11ee-99c3-02420a000123/thumbs/Digital_Inline_Green_400x400.png",
-        label="Instantly connect WhatsApp via a test number, connect your own or buy a number on us. [Help Guide](https://gooey.ai/docs/)",
-    ),
-    ConnectChoice(
-        platform=Platform.TWILIO,
-        img="https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/b0830a1e-0b8c-11f1-a876-02420a000176/Adobe%20Express%20-%20file.png",
-        label="Call or text your copilot with a free test number (or buy one).",
-    ),
-    ConnectChoice(
-        platform=Platform.SLACK,
-        img="https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/ee8c5b1c-d6c8-11ee-b278-02420a000126/thumbs/image_400x400.png",
-        label="Connect to a Slack Channel. [Help Guide](https://gooey.ai/docs/guides/copilot/deploy-to-slack)",
-    ),
-    ConnectChoice(
-        platform=Platform.FACEBOOK,
-        img="https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/9f201a92-1e9d-11ef-884b-02420a000134/thumbs/image_400x400.png",
-        label="Connect to a Facebook Page you own. [Help Guide](https://gooey.ai/docs/guides/copilot/deploy-to-facebook)",
-    ),
-    ConnectChoice(
-        platform=Platform.TELEGRAM,
-        img="https://storage.googleapis.com/dara-c1b52.appspot.com/daras_ai/media/40df4ffa-29d8-11f1-94de-02420a00014b/telegram_logo.png",
-        label="Connect to Telegram.",
-    ),
-]
